Deck in and out
Have the deck slide in and out from the bottom, disabled scrolling, added videos and shows to rotation, handle images for categories that don't return images

Should probably just use image urls from earlier results if no results instead of trying again - although that will usually just be a default image, which would be boring if it repeats a lt.  On the other hand, sometimes a result can hang a while if the next set of results all come back with no images. Can make a decision when I get to caching or retrieving results in advance