Change in getting load screens
Instead of randomly choosing a category then randomly choosing an item from there, get the total number of possible results, generate a random number based on that, then figure out which category that belongs to and get the appropriate result.

While getting the total number of results - which can take a while currently  - we'll still get screens the old way so we can display them.  Once the categories are filled out we'll have fewer calls to GiantBomb and should weight what's being displayed more to categories that have more items, hopefully reducing the repetition in the smaller categories.